Gives time on dialysis in months

Returns a value for time on dialysis in months by blood group and cPRA.

Usage

dial(
  hiper = TRUE,
  bg = "O",
  mean.dial1 = 35,
  mean.dial2 = 70,
  mean.dial3 = 85,
  sd.dial = 20,
  seed.number = NA
)

Arguments

hiper

A logical value for hipersensitized patients (cPRA > 85%).

bg

A character value for blood group.

mean.dial1

time on dialysis mean value for patients that are neither Hipersensitized nor blood group O, used on Normal distribution for value generation.

mean.dial2

time on dialysis mean value for patients that are Hipersensitized or blood group O, used on Normal distribution for value generation.

mean.dial3

time on dialysis mean value for patients that are Hipersensitized and blood group O, used on Normal distribution for value generation.

sd.dial

time on dialysis standard deviation, used on Normal distribution for value generation.

seed.number

Seed for pseudo random number generator. When NA no seed is applicable.

Value

A value from a normal distribution.

Examples

dial(hiper = TRUE, bg = 'O',
mean.dial1 = 35, mean.dial2 = 70, mean.dial3 = 85,
seed.number = NA)
#> [1] 92